content updated

+++ src/components/pages/Login/ManageALLOrders/ManageAllOrders.js
@@ -11,16 +11,20 @@ const ManageAllOrders = () => {
     }, [])
 
     const handleOrderDelete = id => {
-        const proceed = window.confirm('Are you Sure! you want to detete the Order?')
-        fetch(`https://quiet-fjord-24073.herokuapp.com/orders/${id}`, {
+        const proceed = window.confirm('Are you Sure! you want to delete the Order?')
+        if(proceed){
+            fetch(`https://quiet-fjord-24073.herokuapp.com/orders/${id}`, {
             method: 'DELETE'
         })
         .then(res => res.json())
         .then(result => {
             if(result.deletedCount > 0){
                 alert('Order deleted Successfully.')
+                const restOfOrders = orders.filter(order => order._id !== id)
+                setOrders(restOfOrders);
             }
         })
+        }
     }
     return (
         <div className="container py-5" style={{minHeight: '100vh'}}>
@@ -33,6 +37,7 @@ const ManageAllOrders = () => {
                             <th>Name</th>
                             <th>Email</th>
                             <th>Package</th>
+                            <th>Status</th>
                             <th>Manage Order</th>
                           </tr>
                         </thead>
@@ -42,6 +47,7 @@ const ManageAllOrders = () => {
                                   <td>{order.name}</td>
                                   <td>{order.email}</td>
                                   <td>{order.packageName}</td>
+                                  <th>{order.status}</th>
                                   <td><button onClick={() => handleOrderDelete(order._id)} className="border border-info bg-danger text-light rounded">Delete</button></td>
                                 </tr>
                               </tbody>)
